backend/requirements.txt
# Backend Dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
websockets>=12.0

# LLM Providers
# google-generativeai>=0.3.2 (Deprecated)
google-genai>=0.3.0
httpx>=0.25.2

# Vector Memory (FAISS - Facebook AI Similarity Search)
faiss-cpu>=1.7.0

# LangChain for tracing
langchain>=0.1.0
langsmith>=0.0.87

# Utilities
pydantic>=2.5.3
aiofiles>=23.0.0
numpy>=1.24.0
